Fixup data visualization
